requirements.txt
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
isodate==0.7.2
myjdapi==1.1.10
oauthlib==3.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.4.0
